# src/Sonarr.Api.V5/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Sonarr",
    "description": "Sonarr API docs - The v5 API docs apply to Sonarr v5 only.",
    "license": {
      "name": "GPL-3.0",
      "url": "https://github.com/Sonarr/Sonarr/blob/develop/LICENSE"
    },
    "version": "5.0.0"
  },
  "servers": [
    {
      "url": "{protocol}://{hostpath}",
      "variables": {
        "protocol": {
          "default": "http",
          "enum": [
            "http",
            "https"
          ]
        },
        "hostpath": {
          "default": "localhost:8989"
        }
      }
    }
  ],
  "paths": {
...
    "/api/v5/episode": {
      "get": {
        "tags": [
          "Episode"
        ],
        "parameters": [
          {
            "name": "seriesId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "seasonNumber",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "episodeIds",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int32"
              }
            }
          },
          {
            "name": "episodeFileId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "includeSubresources",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/EpisodeSubresource"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EpisodeResource"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v5/history/since": {
      "get": {
        "tags": [
          "History"
        ],
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "eventType",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/EpisodeHistoryEventType"
            }
          },
          {
            "name": "includeSubresources",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/HistorySubresource"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HistoryResource"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v5/history/series": {
      "get": {
        "tags": [
          "History"
        ],
        "parameters": [
          {
            "name": "seriesId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "eventType",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/EpisodeHistoryEventType"
            }
          },
          {
            "name": "includeSubresources",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/HistorySubresource"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HistoryResource"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v5/history/season": {
      "get": {
        "tags": [
          "History"
        ],
        "parameters": [
          {
            "name": "seriesId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "seasonNumber",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "eventType",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/EpisodeHistoryEventType"
            }
          },
          {
            "name": "includeSubresources",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/HistorySubresource"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HistoryResource"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v5/history/episode": {
      "get": {
        "tags": [
          "History"
        ],
        "parameters": [
          {
            "name": "episodeId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "eventType",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/EpisodeHistoryEventType"
            }
          },
          {
            "name": "includeSubresources",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/HistorySubresource"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HistoryResource"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v5/history/failed/{id}": {
      "post": {
        "tags": [
          "History"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v5/series": {
      "get": {
        "tags": [
          "Series"
        ],
        "parameters": [
          {
            "name": "tvdbId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "includeSubresources",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/SeriesSubresource"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SeriesResource"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Series"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SeriesResource"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeriesResource"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddSeriesOptions": {
        "type": "object",
        "properties": {
          "ignoreEpisodesWithFiles": {
            "type": "boolean"
          },
          "ignoreEpisodesWithoutFiles": {
            "type": "boolean"
          },
          "monitor": {
            "$ref": "#/components/schemas/MonitorTypes"
          },
          "searchForMissingEpisodes": {
            "type": "boolean"
          },
          "searchForCutoffUnmetEpisodes": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "AlternateTitleResource": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "nullable": true
          },
          "seasonNumber": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "sceneSeasonNumber": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "sceneOrigin": {
            "type": "string",
            "nullable": true
          },
          "comment": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CustomFormatResource": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "includeCustomFormatWhenRenaming": {
            "type": "boolean",
            "nullable": true
          },
          "specifications": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CustomFormatSpecificationSchema"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "EpisodeFileResource": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "seriesId": {
            "type": "integer",
            "format": "int32"
          },
          "seasonNumber": {
            "type": "integer",
            "format": "int32"
          },
          "relativePath": {
            "type": "string",
            "nullable": true
          },
          "path": {
            "type": "string",
            "nullable": true
          },
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "dateAdded": {
            "type": "string",
            "format": "date-time"
          },
          "sceneName": {
            "type": "string",
            "nullable": true
          },
          "releaseGroup": {
            "type": "string",
            "nullable": true
          },
          "languages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Language"
            },
            "nullable": true
          },
          "quality": {
            "$ref": "#/components/schemas/QualityModel"
          },
          "customFormats": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CustomFormatResource"
            },
            "nullable": true
          },
          "customFormatScore": {
            "type": "integer",
            "format": "int32"
          },
          "indexerFlags": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "releaseType": {
            "$ref": "#/components/schemas/ReleaseType"
          },
          "mediaInfo": {
            "$ref": "#/components/schemas/MediaInfoResource"
          },
          "qualityCutoffNotMet": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "EpisodeHistoryEventType": {
        "enum": [
          "unknown",
          "grabbed",
          "seriesFolderImported",
          "downloadFolderImported",
          "downloadFailed",
          "episodeFileDeleted",
          "episodeFileRenamed",
          "downloadIgnored"
        ],
        "type": "string"
      },
      "EpisodeResource": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "seriesId": {
            "type": "integer",
            "format": "int32"
          },
          "tvdbId": {
            "type": "integer",
            "format": "int32"
          },
          "episodeFileId": {
            "type": "integer",
            "format": "int32"
          },
          "seasonNumber": {
            "type": "integer",
            "format": "int32"
          },
          "episodeNumber": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "airDate": {
            "type": "string",
            "nullable": true
          },
          "airDateUtc": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "lastSearchTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "runtime": {
            "type": "integer",
            "format": "int32"
          },
          "finaleType": {
            "type": "string",
            "nullable": true
          },
          "overview": {
            "type": "string",
            "nullable": true
          },
          "episodeFile": {
            "$ref": "#/components/schemas/EpisodeFileResource"
          },
          "hasFile": {
            "type": "boolean"
          },
          "monitored": {
            "type": "boolean"
          },
          "absoluteEpisodeNumber": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "sceneAbsoluteEpisodeNumber": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "sceneEpisodeNumber": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "sceneSeasonNumber": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "unverifiedSceneNumbering": {
            "type": "boolean"
          },
          "endTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "grabDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "series": {
            "$ref": "#/components/schemas/SeriesResource"
          },
          "images": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MediaCover"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "EpisodeSubresource": {
        "enum": [
          "series",
          "episodeFile",
          "images"
        ],
        "type": "string"
      },
...
      "HistoryResource": {
        "required": [
          "customFormats",
          "data",
          "languages",
          "quality",
          "sourceTitle"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "episodeId": {
            "type": "integer",
            "format": "int32"
          },
          "seriesId": {
            "type": "integer",
            "format": "int32"
          },
          "sourceTitle": {
            "type": "string",
            "nullable": true
          },
          "languages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Language"
            },
            "nullable": true
          },
          "quality": {
            "$ref": "#/components/schemas/QualityModel"
          },
          "customFormats": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CustomFormatResource"
            },
            "nullable": true
          },
          "customFormatScore": {
            "type": "integer",
            "format": "int32"
          },
          "qualityCutoffNotMet": {
            "type": "boolean"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "downloadId": {
            "type": "string",
            "nullable": true
          },
          "eventType": {
            "$ref": "#/components/schemas/EpisodeHistoryEventType"
          },
          "data": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "nullable": true
          },
          "episode": {
            "$ref": "#/components/schemas/EpisodeResource"
          },
          "series": {
            "$ref": "#/components/schemas/SeriesResource"
          }
        },
        "additionalProperties": false
      },
...
      "HistorySubresource": {
        "enum": [
          "series",
          "episode"
        ],
        "type": "string"
      },
...
      "Language": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "MediaCover": {
        "type": "object",
        "properties": {
          "coverType": {
            "$ref": "#/components/schemas/MediaCoverTypes"
          },
          "url": {
            "type": "string",
            "nullable": true
          },
          "remoteUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "MediaCoverTypes": {
        "enum": [
          "unknown",
          "poster",
          "banner",
          "fanart",
          "screenshot",
          "headshot",
          "clearlogo"
        ],
        "type": "string"
      },
...
      "NewItemMonitorTypes": {
        "enum": [
          "all",
          "none"
        ],
        "type": "string"
      },
...
      "Quality": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "source": {
            "$ref": "#/components/schemas/QualitySource"
          },
          "resolution": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "QualityModel": {
        "type": "object",
        "properties": {
          "quality": {
            "$ref": "#/components/schemas/Quality"
          },
          "revision": {
            "$ref": "#/components/schemas/Revision"
          }
        },
        "additionalProperties": false
      },
...
      "Ratings": {
        "type": "object",
        "properties": {
          "votes": {
            "type": "integer",
            "format": "int32"
          },
          "value": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
...
      "ReleaseType": {
        "enum": [
          "unknown",
          "singleEpisode",
          "multiEpisode",
          "seasonPack"
        ],
        "type": "string"
      },
...
      "SeasonResource": {
        "type": "object",
        "properties": {
          "seasonNumber": {
            "type": "integer",
            "format": "int32"
          },
          "monitored": {
            "type": "boolean"
          },
          "statistics": {
            "$ref": "#/components/schemas/SeasonStatisticsResource"
          },
          "images": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MediaCover"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "SeriesResource": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "alternateTitles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AlternateTitleResource"
            },
            "nullable": true
          },
          "sortTitle": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/SeriesStatusType"
          },
          "ended": {
            "type": "boolean",
            "readOnly": true
          },
          "profileName": {
            "type": "string",
            "nullable": true
          },
          "overview": {
            "type": "string",
            "nullable": true
          },
          "nextAiring": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "previousAiring": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "network": {
            "type": "string",
            "nullable": true
          },
          "airTime": {
            "type": "string",
            "nullable": true
          },
          "images": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MediaCover"
            },
            "nullable": true
          },
          "originalLanguage": {
            "$ref": "#/components/schemas/Language"
          },
          "remotePoster": {
            "type": "string",
            "nullable": true
          },
          "seasons": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SeasonResource"
            },
            "nullable": true
          },
          "year": {
            "type": "integer",
            "format": "int32"
          },
          "path": {
            "type": "string",
            "nullable": true
          },
          "qualityProfileId": {
            "type": "integer",
            "format": "int32"
          },
          "seasonFolder": {
            "type": "boolean"
          },
          "monitored": {
            "type": "boolean"
          },
          "monitorNewItems": {
            "$ref": "#/components/schemas/NewItemMonitorTypes"
          },
          "useSceneNumbering": {
            "type": "boolean"
          },
          "runtime": {
            "type": "integer",
            "format": "int32"
          },
          "tvdbId": {
            "type": "integer",
            "format": "int32"
          },
          "tvRageId": {
            "type": "integer",
            "format": "int32"
          },
          "tvMazeId": {
            "type": "integer",
            "format": "int32"
          },
          "tmdbId": {
            "type": "integer",
            "format": "int32"
          },
          "malIds": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "aniListIds": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "firstAired": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "lastAired": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "seriesType": {
            "$ref": "#/components/schemas/SeriesTypes"
          },
          "cleanTitle": {
            "type": "string",
            "nullable": true
          },
          "imdbId": {
            "type": "string",
            "nullable": true
          },
          "titleSlug": {
            "type": "string",
            "nullable": true
          },
          "rootFolderPath": {
            "type": "string",
            "nullable": true
          },
          "folder": {
            "type": "string",
            "nullable": true
          },
          "certification": {
            "type": "string",
            "nullable": true
          },
          "genres": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "tags": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "added": {
            "type": "string",
            "format": "date-time"
          },
          "addOptions": {
            "$ref": "#/components/schemas/AddSeriesOptions"
          },
          "ratings": {
            "$ref": "#/components/schemas/Ratings"
          },
          "statistics": {
            "$ref": "#/components/schemas/SeriesStatisticsResource"
          },
          "episodesChanged": {
            "type": "boolean",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "SeriesStatisticsResource": {
        "type": "object",
        "properties": {
          "seasonCount": {
            "type": "integer",
            "format": "int32"
          },
          "episodeFileCount": {
            "type": "integer",
            "format": "int32"
          },
          "episodeCount": {
            "type": "integer",
            "format": "int32"
          },
          "totalEpisodeCount": {
            "type": "integer",
            "format": "int32"
          },
          "monitoredEpisodeCount": {
            "type": "integer",
            "format": "int32"
          },
          "sizeOnDisk": {
            "type": "integer",
            "format": "int64"
          },
          "releaseGroups": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "percentOfEpisodes": {
            "type": "number",
            "format": "double",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "SeriesStatusType": {
        "enum": [
          "continuing",
          "ended",
          "upcoming",
          "deleted"
        ],
        "type": "string"
      },
      "SeriesSubresource": {
        "enum": [
          "seasonImages"
        ],
        "type": "string"
      },
...
      "SeriesTypes": {
        "enum": [
          "standard",
          "daily",
          "anime"
        ],
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "X-Api-Key": {
        "type": "apiKey",
        "description": "Apikey passed as header",
        "name": "X-Api-Key",
        "in": "header"
      },
      "apikey": {
        "type": "apiKey",
        "description": "Apikey passed as query parameter",
        "name": "apikey",
        "in": "query"
      }
    }
  },
  "security": [
    {
      "X-Api-Key": [ ]
    },
    {
      "apikey": [ ]
    }
  ],
...
}